refact: change export method signature
