+++ datacontract/export/avro_idl_converter.py
@@ -87,9 +87,7 @@ avro_primitive_types = set(
 
 
 class AvroIdlExporter(Exporter):
-    def export(self, export_args) -> dict:
-        self.dict_args = export_args
-        data_contract = self.dict_args.get("data_contract")
+    def export(self, data_contract, model, server, sql_server_type, export_args) -> dict:
         return to_avro_idl(data_contract)
