+++ datacontract/export/bigquery_converter.py
@@ -5,18 +5,14 @@ from typing import Dict, List
 from datacontract.model.data_contract_specification import Model, Field, Server
 from datacontract.model.exceptions import DataContractException
 
-from datacontract.export.exporter import Exporter
-from datacontract.utils import _check_models_for_export
+from datacontract.export.exporter import Exporter, _check_models_for_export
 
 
 class BigQueryExporter(Exporter):
-    def export(self, export_args) -> dict:
+    def export(self, data_contract, model, server, sql_server_type, export_args) -> dict:
         self.dict_args = export_args
-        data_contract = self.dict_args.get("data_contract")
-        model_name, model_value = _check_models_for_export(
-            data_contract, self.dict_args.get("model"), self.dict_args.get("export_format")
-        )
-        found_server = data_contract.servers.get(self.dict_args.get("server"))
+        model_name, model_value = _check_models_for_export(data_contract, model, self.export_format)
+        found_server = data_contract.servers.get(server)
         if found_server is None:
             raise RuntimeError(f"Export to bigquery requires selecting a bigquery server from the data contract.")
         if found_server.type != "bigquery":
