+++ datacontract/export/go_converter.py
@@ -5,9 +5,7 @@ from datacontract.export.exporter import Exporter
 
 
 class GoExporter(Exporter):
-    def export(self, export_args) -> dict:
-        self.dict_args = export_args
-        data_contract = self.dict_args.get("data_contract")
+    def export(self, data_contract, model, server, sql_server_type, export_args) -> dict: 
         return to_go_types(data_contract)
