+++ datacontract/export/great_expectations_converter.py
@@ -4,17 +4,12 @@ from typing import Dict, List, Any
 import yaml
 
 from datacontract.model.data_contract_specification import DataContractSpecification, Field, Quality
-from datacontract.export.exporter import Exporter
-from datacontract.utils import _check_models_for_export
+from datacontract.export.exporter import Exporter, _check_models_for_export
 
 
 class GreateExpectationsExporter(Exporter):
-    def export(self, export_args) -> dict:
-        self.dict_args = export_args
-        data_contract = self.dict_args.get("data_contract")
-        model_name, model_value = _check_models_for_export(
-            data_contract, self.dict_args.get("model"), self.dict_args.get("export_format")
-        )
+    def export(self, data_contract, model, server, sql_server_type, export_args) -> dict: 
+        model_name, model_value = _check_models_for_export(data_contract, model, self.export_format)
         return to_great_expectations(
             data_contract,
             model_name,
