+++ datacontract/export/odcs_converter.py
@@ -7,9 +7,7 @@ from datacontract.export.exporter import Exporter
 
 
 class OdcsExporter(Exporter):
-    def export(self, export_args) -> dict:
-        self.dict_args = export_args
-        data_contract = self.dict_args.get("data_contract")
+    def export(self, data_contract, model, server, sql_server_type, export_args) -> dict: 
         return to_odcs_yaml(data_contract)
