+++ datacontract/export/pydantic_converter.py
@@ -6,9 +6,7 @@ from datacontract.export.exporter import Exporter
 
 
 class PydanticExporter(Exporter):
-    def export(self, export_args) -> dict:
-        self.dict_args = export_args
-        data_contract = self.dict_args.get("data_contract")
+    def export(self, data_contract, model, server, sql_server_type, export_args) -> dict: 
         return to_pydantic_model_str(data_contract)
